// src/app/displays/prc_documento_3/interfaces/factor-conversion.interface.ts
export interface FactorConversionInterface {
    factor_Conversion: number;
    producto: number;
    unidad_Medida: number;
    tipo_Factor_Conversion: number;
    tipo_Precio: number;
    factor_Venta: number;
    presentacion: string;
    des_Tipo_Precio: string;
    precio_Unidad: number;
    descuento_Por: any;
    descuento_Val: any;
    fecha_Ini: any;
    fecha_Fin: any;
    nombre_Moneda: string;
    moneda: number;
    des_Tipo_Factor_Conversion: string;
    abreviatura_Tipo_Factor_Conversion: any;
    des_Producto: string;
    descripcion_Alt_Producto: string;
    iD_Producto: string;
    des_Unidad_Medida: string;
    tipo_Precio_Orden: any;
}