// src/app/displays/prc_documento_3/interfaces/precio.interface.ts
export interface PrecioInterface {
    bodega:          number;
    producto:        number;
    unidad_Medida:   number;
    tipo_Precio:     number;
    moneda:          number;
    precio_Unidad:   number;
    descuento_Por:   any;
    des_Tipo_Precio: string;
    moneda_Nombre:   string;
    moneda_ISO_Code: string;
}